file: Adding ioctl and fcntl interfaces to seastar::file
The following interfaces are added to seastar::file class in this commit.
- ioctl : generic ioctl interface for file class; used where ioctl syscall is required
- fcntl : similar usecase where fcntl interface is required to operate with a file

Refer ioctl(2) and fcntl(2) man pages for more details.
The test cases for the above interfaces are also added.

These interfaces are required to enable seastar::file to be used
with emerging devices such as Zoned Devices, NVMe etc, and to enable their
advanced capabilities.

Signed-off-by: Arun George <arun.george@samsung.com>