core: fix hung IOs in fair_queue
When head pointer wraps, the check of _pending.orig_tail
may fail forever, or at least fail until the next time
the counter wraps around.  This manifests as stuck requests
or spikes of high latency.

Add an extra check that allows queues with a _pending
set to proceed if there is available request capacity,
even if their orig_tail would otherwise not permit it.

In the non-wrapped case, available() will be false as
long as any queue has a _pending, preserving the neat
ordering behaviour that the original check gave us.

Signed-off-by: John Spray <jcs@vectorized.io>